+++ parser_library/test/debugging/debugger_test.cpp
@@ -59,7 +59,7 @@ TEST(debugger, stopped_on_entry)
     EXPECT_EQ(std::string_view(sc.item(1).name), "Locals");
     EXPECT_EQ(std::string_view(sc.item(2).name), "Ordinary symbols");
     auto globs = d.variables(sc.item(0).variable_reference);
-    EXPECT_EQ(globs.size(), 8U);
+    EXPECT_EQ(globs.size(), 9U);
     auto locs = d.variables(sc.item(1).variable_reference);
     EXPECT_EQ(locs.size(), 0U);
 
@@ -182,6 +182,7 @@ struct frame_vars
         this->globals["&SYSPARM"];
         this->globals["&SYSSTMT"];
         this->globals["&SYSTEM_ID"];
+        this->globals["&SYSVER"];
     }
     std::unordered_map<std::string, test_var_value> globals;
     std::unordered_map<std::string, test_var_value> locals;